Updated trackRepo to return exceptions in fail cases and changed tests

+++ MusicApi/Services/TrackRepo.cs
@@ -19,11 +19,14 @@ namespace MusicApi.Services
         {
             return await _dbContext.Tracks.FirstOrDefaultAsync(t => t.Id == id);
         }
+        public bool CheckTrackExist(int id)
+        {
+            return _dbContext.Tracks.Any(t => t.Id == id);
+        }
         public async Task<Track> CreateNewTrack(Track newTrack)
         {
             if (newTrack == null || newTrack.Id != 0 || newTrack.Name == null)
-                // throw new ArgumentException("Invalid track data");
-                return null;
+                throw new ArgumentException("Invalid track data");
 
             var createdTrack = await _dbContext.Tracks.AddAsync(newTrack);
             await _dbContext.SaveChangesAsync();
@@ -33,8 +36,10 @@ namespace MusicApi.Services
         public Track UpdateTrack(Track track)
         {
             if (track == null || track.Id == 0 || track.Name == null)
-                // throw new ArgumentException("Invalid track data");
-                return null;
+                throw new ArgumentException("Invalid track data");
+
+            if (!CheckTrackExist(track.Id))
+                throw new ArgumentException("Track does not exist");
 
             var updatedTrack = _dbContext.Tracks.Update(track);
             _dbContext.SaveChanges();
@@ -44,8 +49,10 @@ namespace MusicApi.Services
         public Track DeleteTrack(Track track)
         {
             if (track == null || track.Id == 0 || track.Name == null)
-                // throw new ArgumentException("Invalid track data");
-                return null;
+                throw new ArgumentException("Invalid track data");
+
+            if (!CheckTrackExist(track.Id))
+                throw new ArgumentException("Track does not exist");
 
             var deletedTrack = _dbContext.Tracks.Remove(track);
             _dbContext.SaveChanges();
